+++ nokogumbo-import/test/test_null.rb
@@ -4,7 +4,7 @@ require 'minitest/autorun'
 
 class TestNull < Minitest::Test
   def fragment(s)
-    Nokogiri::HTML5.fragment(s, max_parse_errors: 10)
+    Nokogiri::HTML5.fragment(s, max_errors: 10)
   end
 
   def test_null_char_ref
@@ -83,7 +83,7 @@ class TestNull < Minitest::Test
   def test_doctype_name_states
     # There are two missing here for double quoted PUBLIC and SYSTEM values.
     doc = Nokogiri::HTML5.parse("<!DOCTYPE \u0000\u0000 PUBLIC '\u0000' SYSTEM '\u0000' \u0000>",
-                                max_parse_errors: 10)
+                                max_errors: 10)
     # 12.2.5.54 Before DOCTYPE name state: unexpected-null-character parse
     # error
     # 12.2.5.55 DOCTYPE name state: unexpected-null-character parse error